+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## Release (2026-07-19)
+
+* vite-ember-ssr 0.4.1 (patch)
+
+#### :bug: Bug Fix
+* `vite-ember-ssr`
+  * [#21](https://github.com/evoactivity/vite-ember-ssr/pull/21) feat: add 'getComputedStyle' to browser globals ([@evoactivity](https://github.com/evoactivity))
+
+#### Committers: 1
+- Liam ([@evoactivity](https://github.com/evoactivity))
+
+## Release (2026-07-19)
+
+* vite-ember-ssr 0.4.0 (minor)
+
+#### :rocket: Enhancement
+* `vite-ember-ssr`
+  * [#17](https://github.com/evoactivity/vite-ember-ssr/pull/17) feat: warn loudly when the SSR bundle exports no settled function ([@st-h](https://github.com/st-h))
+
+#### :bug: Bug Fix
+* `vite-ember-ssr`
+  * [#20](https://github.com/evoactivity/vite-ember-ssr/pull/20) feat: add getComputedStyle to global browser constants ([@evoactivity](https://github.com/evoactivity))
+  * [#13](https://github.com/evoactivity/vite-ember-ssr/pull/13) fix: destroy the per-visit ApplicationInstance even when render throws ([@st-h](https://github.com/st-h))
+  * [#15](https://github.com/evoactivity/vite-ember-ssr/pull/15) fix: bind shoebox capture to the render that started the fetch ([@st-h](https://github.com/st-h))
+
+#### Committers: 2
+- Liam ([@evoactivity](https://github.com/evoactivity))
+- Steve ([@st-h](https://github.com/st-h))
+
 ## Release (2026-05-29)
 
 * vite-ember-ssr 0.3.0 (minor)